move alarm methods to alarm dataclass
