+++ pyadtpulse/__init__.py
@@ -316,7 +316,10 @@ class PyADTPulse:
                     # fetch zones first, so that we can have the status
                     # updated with _update_alarm_status
                     await new_site._fetch_zones(None)
-                    new_site._update_alarm_from_soup(soup)
+                    if new_site.alarm_control_panel is not None:
+                        new_site.alarm_control_panel._update_alarm_from_soup(soup)
+                    else:
+                        LOG.error("Could not fetch control panel information")
                     new_site._update_zone_from_soup(soup)
                     with self._attribute_lock:
                         self._site = new_site
