+++ pyadtpulse/site.py
@@ -12,7 +12,8 @@ from warnings import warn
 from bs4 import BeautifulSoup
 from dateutil import relativedelta
 
-from .const import ADT_ARM_DISARM_URI, ADT_DEVICE_URI, ADT_SYSTEM_URI, LOG
+from .alarm_panel import ADTPulseAlarmPanel
+from .const import ADT_DEVICE_URI, ADT_SYSTEM_URI, LOG
 from .pulse_connection import ADTPulseConnection
 from .util import DebugRLock, make_soup, remove_prefix
 from .zones import (
@@ -22,14 +23,6 @@ from .zones import (
     ADTPulseZones,
 )
 
-ADT_ALARM_AWAY = "away"
-ADT_ALARM_HOME = "stay"
-ADT_ALARM_OFF = "off"
-ADT_ALARM_UNKNOWN = "unknown"
-ADT_ALARM_ARMING = "arming"
-ADT_ALARM_DISARMING = "disarming"
-ADT_ARM_DISARM_TIMEOUT = timedelta(seconds=20)
-
 
 class ADTPulseSite:
     """Represents an individual ADT Pulse site."""
@@ -38,11 +31,8 @@ class ADTPulseSite:
         "_adt_service",
         "_id",
         "_name",
-        "_status",
-        "_is_force_armed",
-        "_sat",
         "_last_updated",
-        "_last_arm_disarm",
+        "_alarm_panel",
         "_zones",
         "_site_lock",
     )
@@ -58,16 +48,14 @@ class ADTPulseSite:
         self._adt_service = adt_service
         self._id = site_id
         self._name = name
-        self._status = ADT_ALARM_UNKNOWN
-        self._is_force_armed: bool = False
-        self._sat = ""
-        self._last_updated = self._last_arm_disarm = datetime(1970, 1, 1)
+        self._last_updated = datetime(1970, 1, 1)
         self._zones = ADTPulseZones()
         self._site_lock: Union[RLock, DebugRLock]
         if isinstance(self._adt_service._attribute_lock, DebugRLock):
             self._site_lock = DebugRLock("ADTPulseSite._site_lock")
         else:
             self._site_lock = RLock()
+        self._alarm_panel: Optional[ADTPulseAlarmPanel] = None
 
     @property
     def id(self) -> str:
@@ -89,75 +77,6 @@ class ADTPulseSite:
 
     # FIXME: should this actually return if the alarm is going off!?  How do we
     # return state that shows the site is compromised??
-    @property
-    def status(self) -> str:
-        """Get alarm status.
-
-        Returns:
-            str: the alarm status
-        """
-        with self._site_lock:
-            return self._status
-
-    @property
-    def is_away(self) -> bool:
-        """Return wheter the system is armed away.
-
-        Returns:
-            bool: True if armed away
-        """
-        with self._site_lock:
-            return self._status == ADT_ALARM_AWAY
-
-    @property
-    def is_home(self) -> bool:
-        """Return whether system is armed at home/stay.
-
-        Returns:
-            bool: True if system is armed home/stay
-        """
-        with self._site_lock:
-            return self._status == ADT_ALARM_HOME
-
-    @property
-    def is_disarmed(self) -> bool:
-        """Return whether the system is disarmed.
-
-        Returns:
-            bool: True if the system is disarmed
-        """
-        with self._site_lock:
-            return self._status == ADT_ALARM_OFF
-
-    @property
-    def is_force_armed(self) -> bool:
-        """Return whether the system is armed in bypass mode.
-
-        Returns:
-            bool: True if system armed in bypass mode
-        """
-        with self._site_lock:
-            return self._is_force_armed
-
-    @property
-    def is_arming(self) -> bool:
-        """Return if system is attempting to arm.
-
-        Returns:
-            bool: True if system is attempting to arm
-        """
-        with self._site_lock:
-            return self._status == ADT_ALARM_ARMING
-
-    @property
-    def is_disarming(self) -> bool:
-        """Return if system is attempting to disarm.
-
-        Returns:
-            bool: True if system is attempting to disarm
-        """
-        with self._site_lock:
-            return self._status == ADT_ALARM_DISARMING
 
     @property
     def last_updated(self) -> datetime:
@@ -180,140 +99,45 @@ class ADTPulseSite:
         """
         return self._site_lock
 
-    async def _arm(self, mode: str, force_arm: bool) -> bool:
-        """Set arm status.
+    def arm_home(self, force_arm: bool) -> bool:
+        """Arm system home."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot arm system home, no control panels exist")
+        return self.alarm_control_panel.arm_home(self._adt_service, force_arm=force_arm)
 
-        Args:
-            mode (str)
-            force_arm (bool): True if arm force
-
-        Returns:
-            bool: True if operation successful
-        """
-        LOG.debug(f"Setting ADT alarm '{self._name}' to '{mode}, force = {force_arm}'")
-        if self._status == mode:
-            LOG.warning(
-                f"Attempting to set alarm status {mode} to "
-                f"existing status {self._status}"
-            )
-            return False
-        if self._status != ADT_ALARM_OFF and mode != ADT_ALARM_OFF:
-            LOG.warning(f"Cannot set alarm status from {self._status} to {mode}")
-            return False
-        params = {
-            "href": "rest/adt/ui/client/security/setArmState",
-            "armstate": self._status,  # existing state
-            "arm": mode,  # new state
-            "sat": self._sat,
-        }
-        if force_arm and mode != ADT_ALARM_OFF:
-            params = {
-                "href": "rest/adt/ui/client/security/setForceArm",
-                "armstate": "forcearm",  # existing state
-                "arm": mode,  # new state
-                "sat": self._sat,
-            }
-
-        response = await self._adt_service._async_query(
-            ADT_ARM_DISARM_URI,
-            method="POST",
-            extra_params=params,
-            timeout=10,
-        )
-
-        soup = await make_soup(
-            response,
-            logging.WARNING,
-            f"Failed updating ADT Pulse alarm {self._name} to {mode}",
-        )
-        if soup is None:
-            return False
-
-        arm_result = soup.find("div", {"class": "p_armDisarmWrapper"})
-        if arm_result is not None:
-            error_block = arm_result.find("div")
-            if error_block is not None:
-                error_text = arm_result.get_text().replace("Arm AnywayCancel\n\n", "")
-                LOG.warning(
-                    f"Could not set alarm state to {mode} " f"because {error_text}"
-                )
-                return False
-        with self._site_lock:
-            self._is_force_armed = force_arm
-            if mode == ADT_ALARM_OFF:
-                self._status = ADT_ALARM_DISARMING
-            else:
-                self._status = ADT_ALARM_ARMING
-            self._last_arm_disarm = self._last_updated = datetime.now()
-            return True
-
-    def _sync_set_alarm_mode(self, mode: str, force_arm: bool = False) -> bool:
-        loop = self._adt_service.loop
-        if loop is None:
-            raise RuntimeError(
-                "Attempting to sync change alarm mode from async session"
-            )
-        coro = self._arm(mode, force_arm)
-        return run_coroutine_threadsafe(coro, loop).result()
-
-    def arm_away(self, force_arm: bool = False) -> bool:
-        """Arm the alarm in Away mode.
-
-        Args:
-            force_arm (bool, Optional): force system to arm
-
-        Returns:
-            bool: True if arm succeeded
-        """
-        return self._sync_set_alarm_mode(ADT_ALARM_AWAY, force_arm)
-
-    def arm_home(self, force_arm: bool = False) -> bool:
-        """Arm the alarm in Home mode.
-
-        Args:
-            force_arm (bool, Optional): force system to arm
-
-        Returns:
-            bool: True if arm succeeded
-        """
-        return self._sync_set_alarm_mode(ADT_ALARM_HOME, force_arm)
+    def arm_away(self, force_arm: bool) -> bool:
+        """Arm system away."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot arm system away, no control panels exist")
+        return self.alarm_control_panel.arm_away(self._adt_service, force_arm=force_arm)
 
     def disarm(self) -> bool:
-        """Disarm the alarm.
+        """Disarm system."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot disarm system, no control panels exist")
+        return self.alarm_control_panel.disarm(self._adt_service)
 
-        Returns:
-            bool: True if disarm succeeded
-        """
-        return self._sync_set_alarm_mode(ADT_ALARM_OFF, False)
+    async def async_arm_home(self, force_arm: bool) -> bool:
+        """Arm system home async."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot arm system home, no control panels exist")
+        return await self.alarm_control_panel.async_arm_home(
+            self._adt_service, force_arm=force_arm
+        )
 
-    async def async_arm_away(self, force_arm: bool = False) -> bool:
-        """Arm alarm away async.
-
-        Args:
-            force_arm (bool, Optional): force system to arm
-
-        Returns:
-            bool: True if arm succeeded
-        """
-        return await self._arm(ADT_ALARM_AWAY, force_arm)
-
-    async def async_arm_home(self, force_arm: bool = False) -> bool:
-        """Arm alarm home async.
-
-        Args:
-            force_arm (bool, Optional): force system to arm
-        Returns:
-            bool: True if arm succeeded
-        """
-        return await self._arm(ADT_ALARM_HOME, force_arm)
+    async def async_arm_away(self, force_arm: bool) -> bool:
+        """Arm system away async."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot arm system away, no control panels exist")
+        return await self.alarm_control_panel.async_arm_away(
+            self._adt_service, force_arm=force_arm
+        )
 
     async def async_disarm(self) -> bool:
-        """Disarm alarm async.
-
-        Returns:
-            bool: True if disarm succeeded
-        """
-        return await self._arm(ADT_ALARM_OFF, False)
+        """Disarm system async."""
+        if self.alarm_control_panel is None:
+            raise RuntimeError("Cannot disarm system, no control panels exist")
+        return await self.alarm_control_panel.async_disarm(self._adt_service)
 
     @property
     def zones(self) -> Optional[List[ADTPulseFlattendZone]]:
@@ -340,65 +164,18 @@ class ADTPulseSite:
             return self._zones
 
     @property
+    def alarm_control_panel(self) -> Optional[ADTPulseAlarmPanel]:
+        """Return the alarm panel object for the site.
+
+        Returns:
+            Optional[ADTPulseAlarmPanel]: the alarm panel object
+        """
+        return self._alarm_panel
+
+    @property
     def history(self):
         """Return log of history for this zone (NOT IMPLEMENTED)."""
         raise NotImplementedError
-
-    def _update_alarm_from_soup(self, summary_html_soup: BeautifulSoup) -> None:
-        LOG.debug("Updating alarm status")
-        value = summary_html_soup.find("span", {"class": "p_boldNormalTextLarge"})
-        sat_location = "security_button_0"
-        with self._site_lock:
-            if value:
-                text = value.text
-                last_updated = datetime.now()
-
-                if re.match("Disarmed", text):
-                    if (
-                        self._status != ADT_ALARM_ARMING
-                        or last_updated - self._last_arm_disarm > ADT_ARM_DISARM_TIMEOUT
-                    ):
-                        self._status = ADT_ALARM_OFF
-                        self._last_updated = last_updated
-                elif re.match("Armed Away", text):
-                    if (
-                        self._status != ADT_ALARM_DISARMING
-                        or last_updated - self._last_arm_disarm > ADT_ARM_DISARM_TIMEOUT
-                    ):
-                        self._status = ADT_ALARM_AWAY
-                        self._last_updated = last_updated
-                elif re.match("Armed Stay", text):
-                    if (
-                        self._status != ADT_ALARM_DISARMING
-                        or last_updated - self._last_arm_disarm > ADT_ARM_DISARM_TIMEOUT
-                    ):
-                        self._status = ADT_ALARM_HOME
-                        self._last_updated = last_updated
-                else:
-                    # FIXME: fix when have gateway device
-                    LOG.warning(f"Failed to get alarm status from '{text}'")
-                    # self._adt_service._set_gateway_status(False)
-                    self._status = ADT_ALARM_UNKNOWN
-                    self._last_updated = last_updated
-                    return
-                LOG.debug(f"Alarm status = {self._status}")
-
-            if self._sat == "":
-                sat_button = summary_html_soup.find(
-                    "input", {"type": "button", "id": sat_location}
-                )
-                if sat_button and sat_button.has_attr("onclick"):
-                    on_click = sat_button["onclick"]
-                    match = re.search(r"sat=([a-z0-9\-]+)", on_click)
-                    if match:
-                        self._sat = match.group(1)
-                elif len(self._sat) == 0:
-                    LOG.warning("No sat recorded and was unable extract sat.")
-
-                if len(self._sat) > 0:
-                    LOG.debug("Extracted sat = %s", self._sat)
-                else:
-                    LOG.warning("Unable to extract sat")
 
         #        status_orb = summary_html_soup.find('canvas', {'id': 'ic_orb'})
         #        if status_orb:
